The Current Working "Satobloc" Build
Re-folding the high water mark SAT-O buld into the most recent, SAT/Blockwave build. The former had the most thorough formalisms, extending to SAT-native covariant equations, string theory integration, and the most built-out mathematical apparatus. The Blockwave version was intended to be a cleaner Lagrangian, a more lightweight, operative version of the theory. The fully integrated SATO/Blockwave or Satobloc version is aiming for the best of both worlds, and initial indications are that its particle spectrum, and fundamental constants are hitting very close to the mark.